refactor(menu): reorganise component typing into a single file
