+++ src/components/menu/index.ts
@@ -1,5 +1,5 @@
 export { default as Menu } from "./menu.component";
-export type { MenuProps } from "./menu.component";
+export type { MenuProps } from "./menu.types";
 export { default as MenuItem } from "./menu-item";
 export type { MenuWithIcon, MenuWithChildren } from "./menu-item";
 export { default as MenuDivider } from "./menu-divider";
