+++ src/components/menu/menu-divider/menu-divider.style.ts
@@ -1,7 +1,8 @@
 import styled, { css } from "styled-components";
 import menuConfigVariants from "../menu.config";
 import { MenuDividerProps } from "./menu-divider.component";
-import { MenuType } from "../__internal__/strict-menu.context";
+
+import type { MenuType } from "../menu.types";
 
 const StyledDivider = styled.div<MenuDividerProps & { menuType: MenuType }>`
   cursor: default;
