+++ src/components/menu/menu-full-screen/menu-full-screen.style.ts
@@ -6,9 +6,10 @@ import StyledSearch from "../../search/search.style";
 import StyledIcon from "../../icon/icon.style";
 import StyledButton from "../../button/button.style";
 import menuConfigVariants from "../menu.config";
-import { MenuType } from "../__internal__/strict-menu.context";
 import addFocusStyling from "../../../style/utils/add-focus-styling";
 import { StyledLink } from "../../link/link.style";
+
+import type { MenuType } from "../menu.types";
 
 const StyledMenuFullscreen = styled.div<{
   transitionDuration: number;
