+++ src/components/menu/menu-segment-title/menu-segment-title.style.ts
@@ -2,7 +2,8 @@ import styled, { css } from "styled-components";
 
 import { VariantType } from "../menu-item";
 import menuConfigVariants from "../menu.config";
-import { MenuType } from "../__internal__/strict-menu.context";
+
+import type { MenuType } from "../menu.types";
 
 interface StyledTitleProps {
   variant?: VariantType;
