+++ src/components/menu/menu.stories.tsx
@@ -8,7 +8,6 @@ import Box from "../box";
 import useMediaQuery from "../../hooks/useMediaQuery";
 import Search, { SearchEvent } from "../search";
 import Typography from "../typography";
-import { MenuType } from "./__internal__/strict-menu.context";
 
 import {
   Menu,
@@ -18,6 +17,8 @@ import {
   ScrollableBlock,
   MenuFullscreen,
 } from ".";
+
+import type { MenuType } from "./menu.types";
 
 const styledSystemProps = generateStyledSystemProps(
   {
