+++ src/components/menu/scrollable-block/scrollable-block.style.ts
@@ -1,11 +1,12 @@
 import styled, { css } from "styled-components";
 import StyledMenuItemWrapper from "../menu-item/menu-item.style";
 import menuConfigVariants from "../menu.config";
-import { MenuType } from "../__internal__/strict-menu.context";
 import { VariantType } from "../menu-item";
 import StyledBox from "../../box/box.style";
 import { StyledMenuItem } from "../menu.style";
 import { StyledLink } from "../../link/link.style";
+
+import type { MenuType } from "../menu.types";
 
 interface StyledScrollableBlockProps {
   menuType: MenuType;
